[fix #160789425] : Refactor db connection and refactor tests
